refactored with gradle and testng
