+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Mon Aug 29 13:09:55 CDT 2016
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.13-bin.zip
