+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.http.multipart.max-file-size=92800000KB
+spring.http.multipart.max-request-size=92800000KB
